Assemble object details in concrete classes for hosts/services
