+++ application/controllers/ServiceController.php
@@ -9,15 +9,14 @@ use Icinga\Module\Icingadb\Common\ServiceLinks;
 use Icinga\Module\Icingadb\Model\History;
 use Icinga\Module\Icingadb\Model\Service;
 use Icinga\Module\Icingadb\Web\Controller;
-use Icinga\Module\Icingadb\Widget\Detail\ObjectDetail;
 use Icinga\Module\Icingadb\Widget\Detail\QuickActions;
+use Icinga\Module\Icingadb\Widget\Detail\ServiceDetail;
 use Icinga\Module\Icingadb\Widget\DowntimeList;
 use Icinga\Module\Icingadb\Widget\ItemList\CommentList;
 use Icinga\Module\Icingadb\Widget\ItemList\HistoryList;
 use Icinga\Module\Icingadb\Widget\ServiceList;
 use Icinga\Module\Icingadb\Widget\ShowMore;
 use ipl\Sql\Sql;
-use ipl\Web\Url;
 
 class ServiceController extends Controller
 {
@@ -61,7 +60,7 @@ class ServiceController extends Controller
         $this->addControl((new ServiceList([$this->service]))->setViewMode('minimal'));
         $this->addControl(new QuickActions($this->service));
 
-        $this->addContent(new ObjectDetail($this->service));
+        $this->addContent(new ServiceDetail($this->service));
 
         $this->setAutorefreshInterval(10);
     }
